Add files via upload
This is a program that displays the mouseListener method. It is a GUI, inside the frame is all the mouseListener methods, each of the method has their own label with their own interaction such as mouse clicked, pressed, released, exited and entered. It also change the background of each method's label.a